pyproject.toml
[project]
name = "11labs-mcp"
version = "0.1.0"
description = "Eleven Labs MCP Server"
requires-python = ">=3.11,<3.13"
dependencies = [
    "mcp>=0.1.0",
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "python-dotenv==1.0.1",
    "pydantic>=2.6.1",
    "httpx==0.28.1",
    "elevenlabs>=1.54.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["11labs_mcp"] 